# perf: math: make Int.Size() faster by computation not len(MarshalledBytes)